Put populate eml function into separate file
